# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  models.ReportSchedule:
    properties:
      created_at:
        type: string
      day_of_month:
        description: 1-31 for monthly reports
        type: integer
      day_of_week:
        description: 0-6 for weekly reports
        type: integer
      format:
        description: json, csv, pdf
        type: string
      frequency:
        description: daily, weekly, monthly
        type: string
      id:
        type: string
      last_run:
        type: string
      name:
        type: string
      next_run:
        type: string
      options:
        type: string
      recipients:
        items:
          type: string
        type: array
      tenant_id:
        type: string
      time_of_day:
        description: HH:MM format
        type: string
      type:
        description: user_activity, location, system_health
        type: string
      updated_at:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: A chat system API with real-time messaging, file sharing, and support
    ticket management.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Chat System API
  version: "1.0"
paths:
...
  /api/v1/scheduler/schedules:
    get:
      consumes:
      - application/json
      description: Get all report schedules for the current tenant
      produces:
      - application/json
      responses:
        "200":
          description: List of schedules
          schema:
            items:
              $ref: '#/definitions/models.ReportSchedule'
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      summary: List report schedules
      tags:
      - Scheduler
    post:
      consumes:
      - application/json
      description: Create a new scheduled report
      parameters:
      - description: Schedule details
        in: body
        name: schedule
        required: true
        schema:
          $ref: '#/definitions/models.ReportSchedule'
      produces:
      - application/json
      responses:
        "201":
          description: Schedule created successfully
          schema:
            $ref: '#/definitions/models.ReportSchedule'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden - Feature not available
          schema:
            additionalProperties: true
            type: object
      summary: Create a report schedule
      tags:
      - Scheduler
  /api/v1/scheduler/schedules/{id}:
    delete:
      consumes:
      - application/json
      description: Delete an existing report schedule
      parameters:
      - description: Schedule ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Schedule deleted successfully
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Schedule not found
          schema:
            additionalProperties: true
            type: object
      summary: Delete a report schedule
      tags:
      - Scheduler
    get:
      consumes:
      - application/json
      description: Get details of a specific report schedule
      parameters:
      - description: Schedule ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Schedule details
          schema:
            $ref: '#/definitions/models.ReportSchedule'
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Schedule not found
          schema:
            additionalProperties: true
            type: object
      summary: Get a report schedule
      tags:
      - Scheduler
    put:
      consumes:
      - application/json
      description: Update an existing report schedule
      parameters:
      - description: Schedule ID
        in: path
        name: id
        required: true
        type: string
      - description: Updated schedule details
        in: body
        name: schedule
        required: true
        schema:
          $ref: '#/definitions/models.ReportSchedule'
      produces:
      - application/json
      responses:
        "200":
          description: Schedule updated successfully
          schema:
            $ref: '#/definitions/models.ReportSchedule'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Schedule not found
          schema:
            additionalProperties: true
            type: object
      summary: Update a report schedule
      tags:
      - Scheduler
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...